+++ client/src/components/meetups/MeetupItem/index.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 import { Meetup } from 'types/meetup';
 import { Link } from 'react-router-dom';
 import MeetupSummary from '../MeetupSummary';
-import MeetupRoute from '../MeetupRoute';
+import MeetupRoutePreview from '../MeetupRoutePreview';
 import styles from './MeetupItem.module.scss';
 import classNames from 'classnames/bind';
 
@@ -16,7 +16,7 @@ const MeetupItem = memo(({ meetup }: MeetupItemProp) => (
   <Link to={`/meetups/${meetup.id}`} className={cn('link')}>
     <article className={cn('meetup')}>
       <MeetupSummary {...meetup} />
-      <MeetupRoute courseName={meetup.courseName} path={meetup.path} />
+      <MeetupRoutePreview courseName={meetup.courseName} path={meetup.path} />
     </article>
   </Link>
 ));
